refactor: extract student enrollment logic into a shared component and enable teacher access to parent search
